% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_auth_shared.R
\name{.api_authenticate}
\alias{.api_authenticate}
\title{Authenticate with an API}
\usage{
.api_authenticate(client, cache_key, scopes, force, refresh_token)
}
\arguments{
\item{client}{A Zoom OAuth client created with \code{\link[=zoom_client]{zoom_client()}}.}

\item{cache_key}{If you are authenticating with multiple users using the same
client, use this key to differentiate between those users.}

\item{scopes}{A character vector of allowed scopes, such as "recording:read".}

\item{force}{A logical indicating whether to force a refresh of the token.}

\item{refresh_token}{A refresh token associated with this \code{client}.}
}
\value{
A Zoom \code{\link[httr2:oauth_token]{httr2::oauth_token()}}, invisibly.
}
\description{
Authenticate with an API
}
\keyword{internal}
