% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_auth_shared.R
\name{.api_browse_oauth_app}
\alias{.api_browse_oauth_app}
\title{Visit an OAuth client management page}
\usage{
.api_browse_oauth_app(app_mgmt_url)
}
\arguments{
\item{app_mgmt_url}{The page where the client can be managed.}
}
\value{
The \code{app_mgmt_url}, invisibly.
}
\description{
Visit an OAuth client management page
}
\keyword{internal}
