% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_auth_shared.R
\name{.api_client}
\alias{.api_client}
\title{Construct an OAuth client}
\usage{
.api_client(client_id = "", client_secret = "")
}
\arguments{
\item{client_id}{A Zoom OAuth App client ID. We recommend you save it as an
environment variable, \code{ZOOM_CLIENT_ID}.}

\item{client_secret}{A Zoom OAuth App client secret. We recommend you save it
as an environment variable, \code{ZOOM_CLIENT_SECRET}.}
}
\value{
A Zoom \code{\link[httr2:oauth_client]{httr2::oauth_client()}}.
}
\description{
Construct an OAuth client
}
\keyword{internal}
