% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_auth_shared.R
\name{.api_get_token_noninteractive}
\alias{.api_get_token_noninteractive}
\title{Retrieve an OAuth token if possible}
\usage{
.api_get_token_noninteractive(client, cache_key = NULL, refresh_token = NULL)
}
\arguments{
\item{client}{A Zoom OAuth client created with \code{\link[=zoom_client]{zoom_client()}}.}

\item{cache_key}{If you are authenticating with multiple users using the same
client, use this key to differentiate between those users.}

\item{refresh_token}{A refresh token associated with this \code{client}.}
}
\value{
A Zoom \code{\link[httr2:oauth_token]{httr2::oauth_token()}}, invisibly.
}
\description{
Retrieve an OAuth token if possible
}
\keyword{internal}
