% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_auth_shared.R
\name{.api_req_authenticate}
\alias{.api_req_authenticate}
\title{Authenticate a request with OAuth2.0}
\usage{
.api_req_authenticate(request, client, scopes, cache_disk, cache_key, token)
}
\arguments{
\item{request}{A \code{\link[httr2:request]{httr2::request()}}.}

\item{client}{A Zoom OAuth client created with \code{\link[=zoom_client]{zoom_client()}}.}

\item{scopes}{A character vector of allowed scopes, such as "recording:read".}

\item{cache_disk}{Deprecated; no longer used. Token caching is handled
in-memory via the package environment. Kept for backward compatibility.}

\item{cache_key}{If you are authenticating with multiple users using the same
client, use this key to differentiate between those users.}

\item{token}{A Zoom API OAuth token, or the \code{access_token} string from such a
token. We recommend that you instead supply a \code{client}, in which case an
appropriate token will be located if possible.}
}
\value{
An \code{\link[httr2:request]{httr2::request()}} with Zoom OAuth authentication information.
}
\description{
Authenticate a request with OAuth2.0
}
\keyword{internal}
