% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_auth_shared.R
\name{.api_token_is_expired}
\alias{.api_token_is_expired}
\title{Is a token expired?}
\usage{
.api_token_is_expired(token)
}
\arguments{
\item{token}{An Zoom API \code{\link[httr2:oauth_token]{httr2::oauth_token()}}.}
}
\value{
A logical scalar indicating whether the token is expired.
}
\description{
Is a token expired?
}
\keyword{internal}
