% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_utils_shared.R
\name{.chr2csv}
\alias{.chr2csv}
\title{Character vector to comma-separated values}
\usage{
.chr2csv(x)
}
\arguments{
\item{x}{A character vector to collapse.}
}
\value{
A character scalar like "this,that".
}
\description{
Collapse a character vector to comma-separate values.
}
\keyword{internal}
