% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_auth_config.R
\name{.oauth-req-auth}
\alias{.oauth-req-auth}
\title{OAuth request authentication documentation}
\arguments{
\item{token}{A Zoom API OAuth token, or the \code{access_token} string from such a
token. We recommend that you instead supply a \code{client}, in which case an
appropriate token will be located if possible.}
}
\value{
An \code{\link[httr2:request]{httr2::request()}} with Zoom OAuth authentication information.
}
\description{
OAuth request authentication documentation
}
