% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_auth_shared.R
\name{.refresh_oauth_token}
\alias{.refresh_oauth_token}
\title{Find and use a refresh token}
\usage{
.refresh_oauth_token(client, cache_key = NULL, refresh_token = NULL)
}
\arguments{
\item{client}{A Zoom OAuth client created with \code{\link[=zoom_client]{zoom_client()}}.}

\item{cache_key}{If you are authenticating with multiple users using the same
client, use this key to differentiate between those users.}

\item{refresh_token}{A refresh token associated with this \code{client}.}
}
\value{
A Zoom \code{\link[httr2:oauth_token]{httr2::oauth_token()}}, invisibly.
}
\description{
Find and use a refresh token
}
\keyword{internal}
