% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_authentication.R
\name{.str2csv}
\alias{.str2csv}
\title{Smush a string to a comma-separated list}
\usage{
.str2csv(string)
}
\arguments{
\item{string}{A character vector to smush.}
}
\value{
A character scalar like "this,that".
}
\description{
Smush a string to a comma-separated list
}
\keyword{internal}
