% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_authentication.R
\name{.zoom_req_auth}
\alias{.zoom_req_auth}
\title{Zoom OAuth authentication}
\usage{
.zoom_req_auth(
  request,
  client = zoom_client(),
  scopes = "recording:read",
  cache_disk = getOption("zoom.cache_disk", FALSE),
  cache_key = getOption("zoom.cache_key", NULL),
  token = NULL
)
}
\arguments{
\item{request}{A \code{\link[httr2:request]{httr2::request()}}.}

\item{client}{A Zoom OAuth client created with \code{\link[=zoom_client]{zoom_client()}}.}

\item{scopes}{A character vector of allowed scopes, such as "recording:read".}

\item{cache_disk}{Should the access token be cached on disk? This reduces
the number of times that you need to re-authenticate at the cost of
storing access credentials on disk. Cached tokens are encrypted and
automatically deleted 30 days after creation.}

\item{cache_key}{If you are authenticating with multiple users using the same
client, use this key to differentiate between those users.}

\item{token}{A Zoom API OAuth token, or the \code{access_token} string from such a
token. We recommend that you instead supply a \code{client}, in which case an
appropriate token will be located if possible.}
}
\value{
A \code{\link[httr2:request]{httr2::request()}} with oauth authentication information.
}
\description{
This function is the main way authentication is handled in this package. It
tries to find a token non-interactively if possible, but bothers the user if
necessary.
}
\keyword{internal}
