% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_auth_zoom.R
\name{.zoom_req_authenticate}
\alias{.zoom_req_authenticate}
\title{Zoom request OAuth authentication}
\usage{
.zoom_req_authenticate(
  request,
  client = zoom_client(),
  scopes = "recording:read",
  cache_disk = getOption("zoom.cache_disk", FALSE),
  cache_key = getOption("zoom.cache_key", NULL),
  token = NULL
)
}
\arguments{
\item{request}{A \code{\link[httr2:request]{httr2::request()}}.}

\item{client}{A Zoom OAuth client created with \code{\link[=zoom_client]{zoom_client()}}.}

\item{scopes}{A character vector of allowed scopes, such as "recording:read".}

\item{cache_disk}{Deprecated; no longer used. Token caching is handled
in-memory via the package environment. Kept for backward compatibility.}

\item{cache_key}{If you are authenticating with multiple users using the same
client, use this key to differentiate between those users.}

\item{token}{A Zoom API OAuth token, or the \code{access_token} string from such a
token. We recommend that you instead supply a \code{client}, in which case an
appropriate token will be located if possible.}
}
\value{
An \code{\link[httr2:request]{httr2::request()}} with Zoom OAuth authentication information.
}
\description{
This function is the main way authentication is handled in this package. It
tries to find a token non-interactively if possible, but bothers the user if
necessary.
}
\keyword{internal}
