% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_auth_zoom.R
\name{zoom_authenticate}
\alias{zoom_authenticate}
\title{Authenticate with a Zoom OAuth client}
\usage{
zoom_authenticate(
  client = zoom_client(),
  cache_key = getOption("zoom.cache_key", NULL),
  scopes = "recording:read",
  force = FALSE,
  refresh_token = NULL
)
}
\arguments{
\item{client}{A Zoom OAuth client created with \code{\link[=zoom_client]{zoom_client()}}.}

\item{cache_key}{If you are authenticating with multiple users using the same
client, use this key to differentiate between those users.}

\item{scopes}{A character vector of allowed scopes, such as "recording:read".}

\item{force}{A logical indicating whether to force a refresh of the token.}

\item{refresh_token}{A refresh token associated with this \code{client}.}
}
\value{
A Zoom \code{\link[httr2:oauth_token]{httr2::oauth_token()}}, invisibly.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#questioning}{\figure{lifecycle-questioning.svg}{options: alt='[Questioning]'}}}{\strong{[Questioning]}}

Load or generate a Zoom OAuth token for use in the other functions in this
package. The primary use of this function is to cache values early in a
script (so the user can walk away). Otherwise the other functions in this
package will prompt for authentication when needed. Once the values are
cached, the rest of this package will use them by default for that client.
}
