% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_authentication.R
\name{zoom_client}
\alias{zoom_client}
\title{Construct a Zoom OAuth client}
\usage{
zoom_client(
  client_id = Sys.getenv("ZOOM_CLIENT_ID"),
  client_secret = Sys.getenv("ZOOM_CLIENT_SECRET")
)
}
\arguments{
\item{client_id}{A Zoom OAuth App client ID. We recommend you save it as an
environment variable, \code{ZOOM_CLIENT_ID}.}

\item{client_secret}{A Zoom OAuth App client secret. We recommend you save it
as an environment variable, \code{ZOOM_CLIENT_SECRET}.}
}
\value{
An \code{\link[httr2:oauth_client]{httr2::oauth_client()}} object.
}
\description{
Builds the OAuth client object for Zoom apis.
}
\examples{
zoom_client <- zoom_client()
}
