% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoomer-package.R
\docType{package}
\name{zoomer-package}
\alias{zoomer}
\alias{zoomer-package}
\title{zoomer: Zoom API in R}
\description{
A consistent interface to the Zoom API, to allow management of Zoom conferencing from R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jonthegeek/zoomer}
  \item \url{https://jonthegeek.github.io/zoomer/}
  \item Report bugs at \url{https://github.com/jonthegeek/zoomer/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})

Authors:
\itemize{
  \item Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})
}

}
\keyword{internal}
