% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_authentication.R
\name{.hack_refresh}
\alias{.hack_refresh}
\title{Hack oauth_flow_refresh for revolving refresh_token}
\usage{
.hack_refresh(
  client = zoom_client(),
  refresh_token = Sys.getenv("ZOOM_REFRESH_TOKEN")
)
}
\arguments{
\item{client}{A Zoom OAuth client created with \code{\link[=zoom_client]{zoom_client()}}.}

\item{refresh_token}{A refresh token. The default behavior is to look in a
\code{ZOOM_REFRESH_TOKEN} environment variable.}
}
\value{
An \code{\link[httr2:oauth_token]{httr2::oauth_token()}}.
}
\description{
This is a (hopefully) temporary hack to trick \code{\link[httr2:oauth_flow_refresh]{httr2::oauth_flow_refresh()}}
into allowing us to refresh even when the refresh token changes.
}
\keyword{internal}
