% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_auth_config.R
\name{.oauth-parameters}
\alias{.oauth-parameters}
\title{OAuth parameters}
\arguments{
\item{cache_disk}{Should the access token be cached on disk? Cached tokens
are encrypted and automatically deleted 30 days after creation. See
\code{\link[httr2:req_oauth_auth_code]{httr2::req_oauth_auth_code()}}.}

\item{cache_key}{If you are authenticating with multiple users using the same
client, use this key to differentiate between those users.}

\item{client}{A Zoom OAuth client created with \code{\link[=zoom_client]{zoom_client()}}.}

\item{client_id}{A Zoom OAuth App client ID. We recommend you save it as an
environment variable, \code{ZOOM_CLIENT_ID}.}

\item{client_secret}{A Zoom OAuth App client secret. We recommend you save it
as an environment variable, \code{ZOOM_CLIENT_SECRET}.}

\item{force}{A logical indicating whether to force a refresh of the token.}

\item{refresh_token}{A refresh token associated with this \code{client}.}

\item{request}{A \code{\link[httr2:request]{httr2::request()}}.}

\item{scopes}{A character vector of allowed scopes, such as "recording:read".}

\item{token}{An Zoom API \code{\link[httr2:oauth_token]{httr2::oauth_token()}}.}
}
\description{
These parameters are used in multiple authentication functions. Define them
here so they're consistent.
}
\keyword{internal}
