% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_authentication.R
\name{zoom_authenticate}
\alias{zoom_authenticate}
\title{Authenticate with a Zoom OAuth client}
\usage{
zoom_authenticate(
  client = zoom_client(),
  scopes = "recording:read",
  force = FALSE,
  refresh_token = NULL
)
}
\arguments{
\item{client}{A Zoom OAuth client created with \code{\link[=zoom_client]{zoom_client()}}.}

\item{scopes}{A character vector of allowed scopes, such as "recording:read".}

\item{force}{A logical indicating whether to force a refresh of the token.}

\item{refresh_token}{A refresh token associated with this \code{client}. This
parameter exists primarily for testing. If you wish to provide a refresh
token (for example, for automated processes), we recommend setting a
\code{ZOOM_REFRESH_TOKEN} environment variable.}
}
\value{
A Zoom OAuth token, invisibly.
}
\description{
Load or generate a Zoom OAuth token for use in the other functions in
this package. The primary use of this function is to cache values early in a
script (so the user can walk away). Otherwise the other functions in this
package will prompt for authentication when needed. Once the values are
cached, the rest of this package will use them by default for that client.
}
