% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_auth_zoom.R
\name{zoom_browse_app_management}
\alias{zoom_browse_app_management}
\title{Visit Zoom Created Apps}
\usage{
zoom_browse_app_management()
}
\value{
The url of the "Created Apps" page, invisibly.
}
\description{
Launch the Zoom "Created Apps" page (where you can configure OAuth 2.0
Clients). Navigate to "Develop > Build App" if you don't have an app, and
choose "OAuth". Set the "OAuth allow list" to
"http://localhost:8888/authorize/". You'll also have to give the app allowed
scopes. After you configure your app, copy/paste the Client ID and Client
Secret into the \code{ZOOM_CLIENT_ID} and \code{ZOOM_CLIENT_ID} environment variables.
We recommend placing these environment variables in your \code{.Renviron} file.
}
\examples{
zoom_app_mgmt_url <- zoom_browse_app_management()
# Copy/paste values from your client.
Sys.setenv(ZOOM_CLIENT_ID = "raNdOMletTeRS")
Sys.setenv(ZOOM_CLIENT_SECRET = "RanDomleTTerSandNumb3rs")
}
